+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed colours: deep sleep is now dark blue, light sleep is now light blue
 * Support for exporting and importing of preferences in addition to the database
 * Visual improvements of the pie charts
+* Add filter by name in the App blacklist activity
 * Pebble: improve compatibility with watch app configuration pages
 * HPlus: users can now decide whether they want to pair the device or not, hopefully fixing some connection problems (#642)
 * HPlus: display battery state and warn on low battery